requirements.txt
requests
black
thop

matplotlib>=3.2.2
scipy>=1.4.1
torchvision>=0.8.1
tensorboard>=2.4.1
seaborn>=0.11.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
numpy==1.21.2
opencv-python==4.5.3.56
pandas==1.3.4
Pillow==8.4.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.25
torch==1.9.1
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
